Only signal when the incompatability set changes
Currently, all waiters are signaled on an unlock and race for the lock.
However, in a case with many pending threads, the set of valid states
won't have changed (say, we've unlocked an IX and gone from 5 holders
to a mere four).  This patch only signals when the count for a particular
context has gone to zero, opening up the possibility that some subset of
waiting threads can now proceed.

Interestingly we didn't see any significant throughput increases in the
benchmarks, but it's not an incorrect patch so it may as well go in.