Remove atomics from state word access
Access to the state word can't be done entirely in a lock-free
manner as we still need invariants maintained by the condvar,
and per #1 it's possible that we'll lose wakeups if it's not
held on the unlock path.  So, remove all atomic accesses to the
state word and ensure the lock is reliably held instead.